requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0

# Visualization and Plotting
matplotlib>=3.5.0
plotly>=5.0.0
seaborn>=0.11.0

# Web UI Framework
streamlit>=1.25.0
streamlit-plotly>=0.0.1

# Data Processing
pandas>=1.3.0
Pillow>=8.3.0
scikit-learn>=1.0.0

# Database
sqlite3

# Configuration Management
PyYAML>=6.0
dataclasses-json>=0.5.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Optional: GPU acceleration
# torch-audio>=2.0.0  # Uncomment if audio processing needed
# torchaudio>=2.0.0   # Uncomment if audio processing needed

# Optional: Advanced visualization
# tensorboard>=2.8.0  # Uncomment for TensorBoard support
# wandb>=0.13.0       # Uncomment for Weights & Biases integration
